+++ packages/schematics/migrations/20180121-update-testjs.ts
@@ -1,3 +1,11 @@
+import * as fs from 'fs';
+
+export default {
+  description: 'Update test.js',
+  run: () => {
+    fs.writeFileSync(
+      'test.js',
+      `
 // This file is required by karma.conf.js and loads recursively all the .spec and framework files
 require('zone.js/dist/zone-testing');
 const getTestBed  = require('@angular/core/testing').getTestBed;
@@ -13,13 +21,17 @@ getTestBed().initTestEnvironment(
   platformBrowserDynamicTesting()
 );
 // Then we find all the tests.
-const contextApps = require.context('./apps', true, /\.spec\.ts$/);
+const contextApps = require.context('./apps', true, /\\.spec\\.ts$/);
 // And load the modules.
 contextApps.keys().map(contextApps);
 
-const contextLibs = require.context('./libs', true, /\.spec\.ts$/);
+const contextLibs = require.context('./libs', true, /\\.spec\\.ts$/);
 // And load the modules.
 contextLibs.keys().map(contextLibs);
 
 // Finally, start Karma to run the tests.
-__karma__.start();
+__karma__.start();    
+    `
+    );
+  }
+};