+++ packages/schematics/src/collection/application/schema.d.ts
@@ -2,8 +2,10 @@ export interface Schema {
   name: string;
   directory: string;
   npmScope?: string;
-  version?: string;
   prefix?: string;
   style?: string;
   minimal?: boolean;
+  skipInstall?: boolean;
+  commit?: { name: string, email: string, message?: string };
+  skipGit?: boolean;
 }