+++ packages/schematics/src/collection/utility/lib-versions.ts
@@ -1,14 +1,18 @@
-export const angularCliVersion = 'file:.angular_cli.tgz';
-export const angularVersion = '5.1.0';
+export const angularCliVersion = '1.7.1';
+export const angularVersion = '5.2.6';
 export const angularJsVersion = '1.6.6';
-export const ngrxVersion = '4.1.1';
+export const ngrxVersion = '5.1.0';
+export const routerStoreVersion = '5.0.1';
 export const nxVersion = '*';
 export const schematicsVersion = '*';
 export const angularCliSchema = './node_modules/@nrwl/schematics/src/schema.json';
-export const latestMigration = '20180116-update-command-line-scripts-to-use-nx-command';
+export const latestMigration = '20180227-cleanup-scripts';
 export const prettierVersion = '1.10.2';
-export const typescriptVersion = '2.5.3';
-export const rxjsVersion = '^5.5.2';
+export const typescriptVersion = '2.6.2';
+export const rxjsVersion = '^5.5.6';
+export const devKitCoreVersion = '^0.0.29';
+export const devKitSchematicsVersion = '0.0.52';
+export const schematicsAngularVersion = '0.1.17';
 
 export const libVersions = {
   angularVersion,
@@ -21,5 +25,9 @@ export const libVersions = {
   angularCliSchema,
   latestMigration,
   typescriptVersion,
-  rxjsVersion
+  rxjsVersion,
+  devKitCoreVersion,
+  devKitSchematicsVersion,
+  schematicsAngularVersion,
+  routerStoreVersion
 };