+++ packages/schematics/src/command-line/affected.ts
@@ -31,9 +31,7 @@ export function affected(args: string[]): void {
 
 function printError(command: string, e: any) {
   console.error(`Pass the SHA range, as follows: npm run affected:${command} -- SHA1 SHA2.`);
-  console.error(
-    `Or pass the list of files, as follows: npm run affected:${command} --files="libs/mylib/index.ts,libs/mylib2/index.ts".`
-  );
+  console.error(`Or pass the list of files, as follows: npm run affected:${command} -- --files="libs/mylib/index.ts,libs/mylib2/index.ts".`);
   console.error(e.message);
 }
 
@@ -41,7 +39,7 @@ function build(apps: string[], rest: string[]) {
   if (apps.length > 0) {
     console.log(`Building ${apps.join(', ')}`);
     apps.forEach(app => {
-      execSync(`${ngPath()} build ${rest.join(' ')} -a=${app}`, { stdio: [0, 1, 2] });
+      execSync(`node ${ngPath()} build ${rest.join(' ')} -a=${app}`, { stdio: [0, 1, 2] });
     });
   } else {
     console.log('No apps to build');
@@ -52,7 +50,7 @@ function e2e(apps: string[], rest: string[]) {
   if (apps.length > 0) {
     console.log(`Testing ${apps.join(', ')}`);
     apps.forEach(app => {
-      execSync(`${ngPath()} e2e ${rest.join(' ')} -a=${app}`, { stdio: [0, 1, 2] });
+      execSync(`node ${ngPath()} e2e ${rest.join(' ')} -a=${app}`, { stdio: [0, 1, 2] });
     });
   } else {
     console.log('No apps to test');
@@ -60,5 +58,6 @@ function e2e(apps: string[], rest: string[]) {
 }
 
 function ngPath() {
-  return `${path.dirname(path.dirname(path.dirname(resolve.sync('@angular/cli', { basedir: __dirname }))))}/bin/ng`;
+  const basePath = path.dirname(path.dirname(path.dirname(resolve.sync('@angular/cli', { basedir: __dirname }))));
+  return path.join(basePath, 'bin', 'ng');
 }